+++ Changelog.md
@@ -1,3 +1,5 @@
+## 0.0.5
+
 ## 0.0.4
 - Disable the port selector while tracing
 - Removed trace DBs for modem firmware 1.2.4 and 1.2.5